reduce pipe buffer + small fixes for processing long execution traces

+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/comptrace/RepetitionMarkerWrapper.java
@@ -42,7 +42,7 @@ public class RepetitionMarkerWrapper implements Serializable {
 		return map;
 	}
 	
-	protected void setTraceSize(int size) {
+	protected void setTraceSize(long size) {
 		this.traceSize = size;
 	}
